message crud

+++ base/views.py
@@ -8,7 +8,7 @@ from django.db.models import Q
 from django.contrib.messages import constants as messages
 from django.contrib import messages
 from django.contrib.auth.forms import UserCreationForm
-from .models import Room, Topic, models
+from .models import Room, Topic, models, Message
 from .form import RoomForm
 
 
@@ -18,6 +18,7 @@ from .form import RoomForm
 #     {'id': 2, 'name': 'Design with me'},
 #     {'id': 3, 'name': 'Backend Developer'}
 # ]
+
 
 
 def loginPage(request):
@@ -45,10 +46,12 @@ def loginPage(request):
     return render(request, 'base/login_register.html', context)
 
 
+
 def logoutUser(request):
     logout(request)
     
     return redirect('home')
+
 
 
 def registerPage(request):
@@ -75,6 +78,7 @@ def registerPage(request):
     return render(request, 'base/login_register.html', context)
 
 
+
 # Perform home page
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''    # equal to paramter passed into the url
@@ -90,7 +94,12 @@ def home(request):
     # get room count
     room_count = rooms.count()
 
-    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
+    room_messages = Message.objects.all()
+
+    print(room_messages)
+
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count,
+                'room_messages': room_messages}
 
     return render(request, 'base/home.html', context)
 
@@ -98,12 +107,30 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=str(pk))
-    context = {'room': room}
+
+    room_messages = room.message_set.all().order_by('-created')     # orderd by newest created date
+
+    participants = room.participants.all()
+
+    print(participants)
+    
+
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body'),
+        )
+        return redirect('room', pk=room.id)
+
+    context = {'room': room, 'room_messages': room_messages,
+                'participants': participants}
 
     return render(request, 'base/room.html', context)
 
 
-# restrice user
+
+# restrict user access
 @login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
@@ -144,9 +171,28 @@ def updateRoom(request, pk):
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
     
+    if request.user != room.host:
+        return HttpResponse('Yout are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     context = {'obj': room}
+
+    return render(request, 'base/delete.html', context)
+
+
+
+@login_required(login_url='login')
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    
+    if request.user != message.user:
+        return HttpResponse('Yout are not allowed here!!')
+
+    if request.method == 'POST':
+        message.delete()
+        return redirect('home')
+    context = {'obj': message}
 
     return render(request, 'base/delete.html', context)
